requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.7
dateparser==1.4.0
fastapi==0.109.2
frozenlist==1.8.0
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gspread==6.0.2
gspread_asyncio==2.0.0
h11==0.16.0
idna==3.13
iniconfig==2.3.0
magic-filter==1.0.12
multidict==6.7.1
oauthlib==3.3.1
packaging==26.1
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
six==1.17.0
starlette==0.36.3
StrEnum==0.4.15
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.46.0
yarl==1.23.0
